Added alternative code in function, fixed typos
Added commented out code. This part is commented out because it isn't needed unless the initial split is something other than 1/4. The initial split was accidentally entered as "tot" rather than 1/4 for these samples, so I hardcoded it to multiply by 4 on line "calculate zooplankton area accounting for 1/4 split". Uncomment if this isn't the case. 
Also fixed comment typos